Added Code Generator scripts
"RCG-Proto.php" generates a random 7 character code, compares the generated code with the codes in "codelist.txt" (not included). If the code does not exist, it adds it to the list, else it will generate a new code and compare again.
"coderedeem-proto.php" takes the code that the user has input and compares it with the codes in "codelist.txt" (not included) to check whether the code does or does not exist. If it does not exist, a message displays that the code does not exist. If it does exist, it then compares the code to the codes in "redeemlist.txt" (not included) to check if it has been previously redeemed. If it does exist, a message displays that the code has already been redeemed. If it does not exist, the code is added to the list.